requirements.txt
# Core LangChain & AI
langchain==1.1.3
langchain-core==1.2.0
langchain-community==0.4.1
langchain-google-genai==1.0.0
langgraph
supabase

# Embeddings
langchain-huggingface==0.0.1
sentence-transformers==3.0.0

# Database
sqlalchemy==2.0.45
psycopg2-binary==2.9.9
pgvector==0.2.4

# Document Processing
pypdf==4.0.1
python-dotenv==1.0.0

# Utilities
pydantic==2.12.5
requests==2.32.5
numpy==2.3.5

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
